### 2018-10-10 74d9fca
References and Acknowledgments Update
-added missing references info to the paper.bib file
-updated Acknowledgements section of paper.md file to acknowledge TEM tomo data from Andy
