### 2018-10-10 3d9f107
Paper Compilation Bugfix
[skip CI]
-updated paper.md to remove potential conflicts with math symbols
