### 2018-10-10 f026d54
Paper Compilation Bugfix
skip ci
-fixed use of underscores in paper.md
