+++ papers/v4_paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Ising_OPV v4.0: Experimental Tomography Data Import, Interpretation, and Analysis'
+title: 'Ising\_OPV v4.0: Experimental Tomography Data Import, Interpretation, and Analysis'
 tags:
   - organic photovoltaics
   - bulk heterojunction morphology
@@ -20,9 +20,9 @@ bibliography: paper.bib
 
 # Summary
 
-Understanding the impact of the complex meso-scale morphology is critical for the development of organic semiconductor materials and devices.  This is particularly important in organic photovoltaics (OPVs), where a blend of two or more components phase separates to form a bulk heterojunction (BHJ) structure.  To build better structure-property models for organic BHJ photovoltaics, the simple Ising-based morphology model has proven to be a highly useful tool when coupled with kinetic Monte Carlo (KMC) simulations.[@heiber2019chapter] ``Ising_OPV`` was originally designed as an efficient, open-source C\texttt{++} tool that would enable researchers in the community to have easy access to this morphology model and allow them to create well-controlled morphologies on an HPC cluster for KMC simulations.[@heiber2014prapp] Demonstrating the utility of this tool, the ability to systematically control the domain size allowed a detailed investigation of the charge carrier recombination kinetics in OPVs.[@heiber2015prl, heiber2016prb] The tool can also create controlled interfacial mixing, which can be important for simulating the exciton dissociation dynamics and charge separation yield in OPVs.[@lyons2012ees, @heiber2013jpcc] In addition, the tool was later updated to add new features that allow further structural control and quantification of important morphological features, most importantly the domain tortuosity.[@heiber2017prapp] The tool has also been used as a testbed for developing more advanced 3D image analysis methods.[@aboulhassan2017cgf]
+Understanding the impact of the complex meso-scale morphology is critical for the development of organic semiconductor materials and devices.  This is particularly important in organic photovoltaics (OPVs), where a blend of two or more components phase separates to form a bulk heterojunction (BHJ) structure.  To build better structure-property models for organic BHJ photovoltaics, the simple Ising-based morphology model has proven to be a highly useful tool when coupled with kinetic Monte Carlo (KMC) simulations.[@heiber2019chapter] ``Ising\_OPV`` was originally designed as an efficient, open-source C\texttt{++} tool that would enable researchers in the community to have easy access to this morphology model and allow them to create well-controlled morphologies on an HPC cluster for KMC simulations.[@heiber2014prapp] Demonstrating the utility of this tool, the ability to systematically control the domain size allowed a detailed investigation of the charge carrier recombination kinetics in OPVs.[@heiber2015prl, heiber2016prb] The tool can also create controlled interfacial mixing, which can be important for simulating the exciton dissociation dynamics and charge separation yield in OPVs.[@lyons2012ees, @heiber2013jpcc] In addition, the tool was later updated to add new features that allow further structural control and quantification of important morphological features, most importantly the domain tortuosity.[@heiber2017prapp] The tool has also been used as a testbed for developing more advanced 3D image analysis methods.[@aboulhassan2017cgf]
 
[…]
