### 2018-10-12 #1011 @MikeHeiber
@arfon @labarba Any idea why the PDF is not compiling? I am not using any math in the document, so I can't figure out why latex is complaining.
