### 2018-10-12 #1011 @arfon
I think the issue is actually with the underscore in the title. This pull
request should fix the paper: https://github.com/MikeHeiber/Ising_OPV/pull/8
