### 2018-10-12 #1011 @MikeHeiber
I tried that as well, and it didn't seem to fix the problem.  See the joss-edits branch for my latest edited version.
