### 2018-10-12 #1011 @arfon
@MikeHeiber - I'm not sure what's going on here. I can compile your paper fine with the edits in your `joss-edits` branch: [10.21105.joss.01011.pdf](https://github.com/openjournals/joss-reviews/files/2473426/10.21105.joss.01011.pdf)
