### 2018-10-12 b098cc3
Paper Update
skip CI
-corrected some formatting errors and typos
-added further explanation to the figure caption
