+++ papers/v4_paper/paper.bib
@@ -2,6 +2,7 @@
 author = {A. Aboulhassan and R. Sicat and D. Baum and O. Wodo and M. Hadwiger},
 title = {Comparative Visual Analysis of Structure-Performance Relations in Complex Bulk-Heterojunction Morphologies},
 journal = {Comput. Graph. Forum},
+year = {2017},
 volume = {36},
 number = {3},
 pages = {329-339},
