### 2018-10-12 #1011 @MikeHeiber
@arfon Hmm... Well thanks for providing the compiled PDF.  Based on how that compiled, I've made a few more minor corrections/modifications to the joss-edits branch.
