### 2018-10-12 #1011 @arfon
Compiled again locally [10.21105.joss.01011.pdf](https://github.com/openjournals/joss-reviews/files/2474390/10.21105.joss.01011.pdf)
