### 2018-10-15 #1011 @MikeHeiber
The proof looks good to me!
