### 2018-10-17 #1011 @labarba
You should expect at least a period of two weeks for finding reviewers. Sometimes more. Editors and reviewers are volunteers and most often have latency before new tasks enter our queue. Right now, for example, I'm starting a trip, with two talks to give, and last week, another trip and another talk.
