### 2018-11-07 #1011 @MikeHeiber
@arfon Thank you for trying to move this along.  With the traditional publishers that I submit to, most of the time is spent waiting for the reviewers to assess the technical aspects of the work and not on waiting for the editors to make initial editorial decisions and sending the manuscript out for review.  I am happy to put in some more work to comb over the reviewer list again and suggest reviewers if it helps get things going.  I could even suggest new JOSS reviewers that work in my field and are active Github users, if it would help.
