### 2018-11-08 #1011 @katyhuff
@MikeHeiber This level of materials science is slightly beyond my wheelhouse, so I spent a few hours on the plane looking over your submission to understand the purpose. At this point, I think I grok it, and I'm certainly happy to help out in finding reviewers unless @labarba has an objection ( @labarba I know you're on travel and juggling more papers than most, but if you object to changing editors I'm happy to return this to your capable hands ).
