### 2018-11-08 #1011 @katyhuff
Now, @MikeHeiber ,  let's ping a few fitting reviewers, with materials science, physics, and c or c++ expertise. We can start the review once two or three of them accept the request to review.

----

Dear reviewers,

- @stuartcampbell as suggested: Stuart, are you able to and interested in taking on this review?
- @mdoucet  Mathieu, are you able to and interested in taking on this review?
- @jochym : Pawel, are you able to and interested in taking on this review?
- @trallard : Tania, are you able to and interested in taking on this review?
- @myousefi2016 : Mehrdad, are you able to and interested in taking on this review?

**Title:** Ising_OPV v4.0: Experimental Tomography Data Import, Interpretation, and Analysis
**Summary:** A C++ software tool for generating and analyzing model bulk heterojunction morphologies in a parallel computing environment
**Topic:** Computational Materials Science
**Article Proof:** https://github.com/openjournals/joss-reviews/files/2474390/10.21105.joss.01011.pdf
**Submitting author:** @MikeHeiber (Michael Heiber)
**Repository:** https://github.com/MikeHeiber/Ising_OPV
**Version:** v4.0.0-rc.2
