### 2018-11-09 #1072 @MikeHeiber
@arfon Could you compile the again locally for me from the joss-edits branch?  Still not sure why the compilation is failing.
