### 2018-11-09 #1072 @arfon
> @arfon Could you compile the again locally for me from the joss-edits branch? Still not sure why the compilation is failing.

Sure thing: [10.21105.joss.01072.pdf](https://github.com/openjournals/joss-reviews/files/2566544/10.21105.joss.01072.pdf)
