### 2018-11-17 MikeHeiber/Ising_OPV#10 @myousefi2016
MPI test failed
I build your code with OpenMPI/1.10.3 and gcc/6.3.0 and then used this command: `./test/Ising_OPV_MPI_tests.exe` but it shows this error:

```
[==========] Running 5 tests from 1 test case.
[----------] Global test environment set-up.
[----------] 5 tests from MPI_Test
[ RUN      ] MPI_Test.CalculateProbHistAvgTests
test/test_mpi.cpp:46: Failure
Expected equality of these values:
  6
  (int)prob.size()
    Which is: 3
[  FAILED  ] MPI_Test.CalculateProbHistAvgTests (0 ms)
[ RUN      ] MPI_Test.CalculateVectorAvgTests
test/test_mpi.cpp:68: Failure
Expected equality of these values:
  4.5
  data_avg[0]
    Which is: 0
test/test_mpi.cpp:69: Failure
Expected equality of these values:
  5.5
  data_avg[1]
    Which is: 1
test/test_mpi.cpp:70: Failure
Expected equality of these values:
  6.5
  data_avg[2]
    Which is: 2
[  FAILED  ] MPI_Test.CalculateVectorAvgTests (1 ms)
[ RUN      ] MPI_Test.CalculateVectorSumTests
test/test_mpi.cpp:83: Failure
Expected equality of these values:
  18.0
    Which is: 18
  data_sum[0]
    Which is: 0
test/test_mpi.cpp:84: Failure
Expected equality of these values:
  22.0
    Which is: 22
  data_sum[1]
    Which is: 1
test/test_mpi.cpp:85: Failure
Expected equality of these values:
  26.0
    Which is: 26
  data_sum[2]
    Which is: 2
test/test_mpi.cpp:94: Failure
Expected equality of these values:
  18
  data_sum2[0]
    Which is: 0
test/test_mpi.cpp:95: Failure
Expected equality of these values:
  22
  data_sum2[1]
    Which is: 1
test/test_mpi.cpp:96: Failure
Expected equality of these values:
  26
  data_sum2[2]
    Which is: 2
[  FAILED  ] MPI_Test.CalculateVectorSumTests (0 ms)
[ RUN      ] MPI_Test.GatherValuesTests
[       OK ] MPI_Test.GatherValuesTests (0 ms)
[ RUN      ] MPI_Test.GatherVectorsTests
test/test_mpi.cpp:151: Failure
Expected equality of these values:
  12
  (int)data_all.size()
    Which is: 3
test/test_mpi.cpp:162: Failure
Expected equality of these values:
  12
  (int)data_all2.size()
    Which is: 3
[  FAILED  ] MPI_Test.GatherVectorsTests (0 ms)
[----------] 5 tests from MPI_Test (1 ms total)

[----------] Global test environment tear-down
[==========] 5 tests from 1 test case ran. (1 ms total)
[  PASSED  ] 1 test.
[  FAILED  ] 4 tests, listed below:
[  FAILED  ] MPI_Test.CalculateProbHistAvgTests
5 FAILED TESTS
```

Any idea how to resolve these errors?
