### 2018-11-18 MikeHeiber/Ising_OPV#10 @MikeHeiber
@myousefi2016  The correct way to perform the mpi tests is using the command:
```mpiexec -n 4 ./test/Ising_OPV_MPI_tests.exe```
from the root dir.  I'll update the README with the correct testing instructions. Let me know if there are any further problems.
