### 2018-11-20 MikeHeiber/Ising_OPV#10 @MikeHeiber
I still need to correct the readme
