### 2018-11-20 #1072 @mdoucet
I've completed my review. Looks good!
