### 2018-11-21 d2706c5
Update README
-corrected mistake in the testing instructions for the mpi tests
-Resolves #10 
[skip ci]
