### 2018-11-21 #1072 @MikeHeiber
@myousefi2016 The [issue you filed](https://github.com/MikeHeiber/Ising_OPV/issues/10) due to incorrect testing instructions in the README has been fixed in the latest version on the [joss-edits branch](https://github.com/MikeHeiber/Ising_OPV/tree/joss-edits).  Once all of the reviews are done I'll merge this branch into master and make a final release v4.0.0 for the paper.
