### 2018-11-28 #1072 @stuartcampbell
Sorry for the delay.

All looks good to me - my review is completed.

One minor suggestion is to consider adding a simple `CONTRIBUTING.md`, which could just link to the sections in your `README.md` and `CODE_OF_CONDUCT.md` - so just to provide a place with a standard name for folks to look.  Not essential for this review - just a personal comment.
