### 2018-11-28 #1072 @katyhuff
Thanks @stuartcampbell , @myousefi2016 ,  @mdoucet for your reviews .

@MikeHeiber thank you for a strong submission and for engaging actively in the review process!
It looks like there are just a couple of minor suggestions from @stuartcampbell you might consider as you make your 4.0.0 release. Once you have your code and paper all ready in the release, it'll be good to double check the paper, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service.  When you update this thread with the DOI of the archive, I'll move forward with accepting the submission! Until then, now is your moment for final touchups!
