### 2018-11-29 862fd07
Create CONTRIBUTING.md
-added some more detailed instructions for how others can contribute to the project
