+++ CONTRIBUTING.md
@@ -1,0 +1,94 @@
+# Contributing to Ising_OPV
+
+We welcome new contributors to this project to help make the Ising_OPV morphology tool better and more effective for our scientific and educational users. Contributions can take on a variety of forms, such as bug reports, feature requests, code documentation updates, user tutorials, feature implementations, bugfixes, and more.
+
+## Ground Rules
+
+- To maintain an inclusive and welcoming environment, all community members must follow the [Code of Conduct](https://github.com/MikeHeiber/Ising_OPV/blob/master/CODE_OF_CONDUCT.md).
+- To keep project development organized, all new features and bugfixes must be proposed and discussed using [issues](https://github.com/MikeHeiber/Ising_OPV/issues) before changes are implemented.
+- C++ code should follow the [C++11 standard](https://en.wikipedia.org/wiki/C%2B%2B11) and should be written to be simple and largely self-documenting with self-descriptive names.
+
+## Filing Issues
+
+One of the easiest ways to contribute is by using the software and submitting bug reports and feature requests as issues in the [Issues section](https://github.com/MikeHeiber/Ising_OPV/issues). 
+Before filing a new issue please check that your issue has not already been submitted. 
+If it has, 'like' the issue and add any additional feedback you have to the issue thread in a comment.
+If your issue has not already been filed, create a new issue and provide a detailed description of the problem or new idea/suggestion. 
+The code owners will review new issues, provide feedback, and decide if and when to implement the changes needed to resolve the issue.
+
+## Working on an Issue
+
+1. Before you set out to tackle an issue, discuss your implementation plans with the code owners on the issue page.
+    
+    Especially for code additions needed to add new features, it is important to discuss your implementation ideas with the code owners to ensure that the software design is relatively well thought out and fits with the rest of the codebase.
+    
+2. Create fork of the Ising_OPV repo on your account.
+
+    Click the "Fork" button near the top of any Ising_OPV Github webpage.
+
+3. Connect your fork to the upstream repo
+
+    ```
+    git remote add upstream https://github.com/MikeHeiber/Ising_OPV.git
+    ```
+
+4. Clone your forked Ising_OPV repo to your local machine.
+
+    ```git clone --recurse-submodules https://github.com/YOUR_USERNAME/Ising_OPV```
+
+5. Create a new feature branch from the development branch.
+
+    ```
+    git checkout development
+    git branch feature-new-thing
+    git checkout feature-new-thing
+    ```
+
+6. Implement your changes keeping the scope of changes limited to the specified issue.
+
+7. Document your changes by adding to the Changelog.
+
+8. Update your new branch with any upstream changes and resolve any conflicts.
+
+    ```
+    git fetch upstream
+    git checkout development
+    git merge upstream/development
+    git checkout feature-new-thing
+    git rebase development
+    ```
+
+9. Make sure your changes do not break any of the existing functionality by running the tests locally.
+
+    ```
+    make test
+    ./test/Ising_OPV_tests.exe
+    mpiexec -n 4 ./test/Ising_OPV_MPI_tests.exe
+    ```
+    
+10. Push your local changes to your forked Ising_OPV repo.
+
+    ```
+    git push origin feature-new-thing
+    ```
+
+11. Submit a pull request
+
+    See below for details about how to submit a pull request.
+
+## Code Testing and Documentation Standards
+
+Unit tests using the [googletest framework](https://github.com/abseil/googletest/blob/master/googletest/docs/primer.md) must be created for all new/modified functions and added to the test/test.cpp or test/test_mpi.cpp file as appropriate. Tests should include comments that describe what functionality each test code section is testing for.
+
+All new/modified classes and functions must be documented using the [Doxygen style](https://www.stack.nl/~dimitri/doxygen/manual/docblocks.html).
+
[…]
