### 2018-11-29 5b6081b
Update README
-updated current status to highlight that v4.0 is now stable
-added link to contributing instructions
