### 2018-11-29 b6702a0
Update README
[skip ci]
-updated current status section
