### 2018-11-29 284663d
Update CONTRUBUTING.md
[skip ci]
-updated link to code of conduct file to use a relative path
