### 2018-11-29 008a3c5
Update CONTRIBUTING.md
[skip ci]
-added link to the changelog
