### 2018-11-29 e510bd7
Update CONTRIBUTING.md
[skip ci]
-improved formatting
