### 2018-11-29 f121e09
Release Version Update
General:
-generated updated Doxygen documentation
-updated Doxyfile version to v4.0.0
-updated Changelog by adding changes in this commit

Version class:
-updated the version string to v4.0.0
