### 2018-11-29 ad6d6db
Joss edits (#12)
- Resolves #9 
- Resolves #10 
- Resolves #11 

General:
-added missing references info to the paper.bib file
-updated Acknowledgements section of paper.md file to acknowledge TEM tomo data from Andy
-updated paper.md to remove potential conflicts with math symbols
-fixed use of underscores in paper.md
-corrected some formatting errors and typos in paper.md
-added further explanation to the figure caption in paper.md
-corrected mistake in the testing instructions for the mpi tests in README
-added CONTRIBUTING.md file with some more detailed instructions for how others can con
[…]
