+++ papers/v4_paper/paper.bib
@@ -2,6 +2,7 @@
 author = {A. Aboulhassan and R. Sicat and D. Baum and O. Wodo and M. Hadwiger},
 title = {Comparative Visual Analysis of Structure-Performance Relations in Complex Bulk-Heterojunction Morphologies},
 journal = {Comput. Graph. Forum},
+year = {2017},
 volume = {36},
 number = {3},
 pages = {329-339},
@@ -72,6 +73,14 @@ howpublished = {\url{https://github.com/MikeHeiber/Excimontec}},
 author = {Michael C. Heiber}
 }
 
+@misc {heiber2018ising4,
+author = "Michael C. Heiber",
+title = "{I}sing\_{OPV} v4.0.0 Release Candidate 2",
+url = "https://doi.org/10.5281/zenodo.1452005",
+year = "2018",
+doi = "10.5281/zenodo.1452005",
+}
+
 @incollection {heiber2019chapter,
 author = "Michael C. Heiber and Alex Wagenpfahl and Carsten Deibel",
 title = "Advances in Modeling the Physics of Disordered Organic Electronic Devices",
@@ -94,4 +103,39 @@ volume = "5",
 pages = "7657-7663",
 doi = "10.1039/C2EE21327C",
 publisher = "The Royal Society of Chemistry",
+}
+
+@article {pfannmoeller2013ees,
+author = "Martin Pfannm{\"o}ller and Wolfgang Kowalsky and Rasmus R. Schr{\"o}der",
+title = "Visualizing Physical, Electronic, and Optical Properties of Organic Photovoltaic Cells",
+year = "2013",
+journal = "Energy Environ. Sci.",
+volume = "6",
+pages = "2871-2891",
+doi = "10.1039/c3ee41773e",
+publisher = "The Royal Society of Chemistry",
+}
+
+@article {proudian2018arXiv,
+author = "Andrew P. Proudian and Matthew B. Jaskot and David R. Diercks and Brian P. Gorman and Jeramy D. Zimmerman",
+title = "Atom Probe Tomography of Organic Molecular Materials: Sub-Dalton Nanometer-Scale Quantification",
+year = "2018",
+journal = "ArXiv e-prints",
+Eprint = "1806.10650",
+volume = "",
+pages = "",
+doi = "",
+publisher = "",
+}
+
+
+@article {vanbavel2009nl,
+author = "Svetlana S. van Bavel and Erwan Sourty and Gijsbertus de With and Joachim Loos",
+title = "Three-Dimensional Nanoscale Organization of Bulk Heterojunction Polymer Solar Cells",
+year = "2009",
+journal = "Nano Lett.",
+volume = "9",
+pages = "507-513",
+doi = "10.1021/nl8014022",
+publisher = "American Chemical Society",
 }
