### 2018-11-29 #1072 @MikeHeiber
@stuartcampbell Thank you for the suggestion about the contributor instructions file.  I had been meaning to add more detailed instructions for how new people could contribute, and your reminder helped give me the activation energy that I needed.  There is now both a CONTRIBUTING.md and a CODE_OF_CONDUCT.md file. In addition, I have also added a CHANGELOG.md file as part of my final improvements.

@myousefi2016 and @mdoucet Thank you for taking the time to review my submission.

@katyhuff All of the changes are now merged into the master branch and are contained in the tagged release [v4.0.0](https://github.com/MikeHeiber/Ising_OPV/releases/tag/v4.0.0) that is archived on [Zenodo](https://doi.org/10.5281/zenodo.1710685) with the following DOI:

10.5281/zenodo.1710685

This DOI needs to be added to the References section of the final paper.

The [Changelog](https://github.com/MikeHeiber/Ising_OPV/blob/master/CHANGELOG.md) lists all changes made between the submitted v4.0.0-rc.2 and the final v4.0.0.

Please let me know if you need any more information from me to move things forward.
