### 2018-11-29 #1072 @katyhuff
@myousefi2016 @mdoucet @stuartcampbell Thank you so much, once again, for your reviews of this work. JOSS relies on reviewers like you!

@MikeHeiber Thanks again for a strong submission, and congratulations, we're ready to accept your paper!

@arfon over to you!
